pyproject.toml
[project]
name = "carsharing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-admin-rangefilter>=0.13.2",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-import-export>=4.3.7",
    "djangorestframework>=3.15.2",
    "drf-standardized-errors>=0.14.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "minio>=7.2.15",
    "openpyxl>=3.1.5",
    "pendulum>=3.0.0",
    "psycopg2-binary>=2.9.10",
    "pytelegrambotapi>=4.26.0",
    "sentry-sdk>=2.22.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "freezegun>=1.5.1",
    "pre-commit>=4.1.0",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.10.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "carsharing.settings"
